requirements.txt
cloudinary==1.32.0
fastapi==0.95.0
pydantic==1.10.2
uvicorn==0.21.1
email-validator==1.3.1
python-multipart==0.0.6
aiofiles==23.1.0
bcrypt==4.0.1
jinja2==3.1.2
fastapi-login==1.9.0
passlib==1.7.4
python-jose==3.4.0